Fix Go keyword highlighting
In the standard Go syntax for vscode, the keywords don't get the same
subscope as in some other languages, such as keyword.other. Instead
non-control keywords all get the scope such as keyword.package.go.

This adds all of the Go keywords as of go 1.24 into their keywords
matcher. The keywords are added separately to avoid conflicts with
other textmate entities with keyword.* scopes, such as
keyword.operator.*.